Move mappers to dedicated files, rethrow CancellationException, rename sync handler
- Move domain-to-infrastructure mapping functions from implementation
  files to dedicated Mapper.kt/Mappers.kt files (BE driven, FE
  driven/api, FE driven/db)
- Rethrow CancellationException in NonWebAddFindingPhotoUseCaseImpl
  to preserve coroutine cancellation propagation
- Rename FindingPhotoSyncHandler to NonWebFindingPhotoSyncHandler
